+++ diffractio/config.py
@@ -30,10 +30,9 @@ CONF_DRAWING['color_phase'] = cm.twilight  # twilight .twilight hsv
 CONF_DRAWING['color_real'] = cm.seismic
 CONF_DRAWING['color_stokes'] = cm.seismic
 CONF_DRAWING['color_n'] = cm.Blues
-# percentage of intensity not shown in phase
-CONF_DRAWING['percentage_intensity'] = 0.0
-# percentage of intensity to show Ez
-CONF_DRAWING['percentage_Ez'] = 0.0
+CONF_DRAWING['percentage_intensity'] = 0.0 # percentage of intensity not shown in phase
+CONF_DRAWING['percentage_Ez'] = 0.0 # percentage of intensity to show Ez
+
 
 
 number_types = (int, float, complex, np.int32, np.float64)
@@ -51,6 +50,7 @@ Draw_Vector_XY_Options = Literal['intensity', 'intensities', 'intensities_rz', '
 Draw_Vector_XZ_Options = Literal[ 'intensity', 'intensities', 'intensities_rz', 'phases', 'fields', 'EH', 'E2H2', 'poynting_vector', 'poynting_vector_averaged', 'poynting_total', 'energy_density', 'irradiance', 'stokes', 'param_ellipses']
 Draw_pyvista_Options = Literal['volume', 'clip', 'slices', 'projections']
 Draw_interactive_Options = Literal['intensity', 'amplitude', 'phase']
+
 Save_mask_Options = Literal['amplitude', 'phase', 'intensity']
 
 Options_Diffractio_kind = Literal['scalar', 'vector']
@@ -58,12 +58,13 @@ Options_Diffractio_frame = Literal['field', 'mask', 'source']
 
 Options_add = Literal['source', 'mask', 'refractive_index', 'phases', 'no_overlap', 'distances']
 Options_sub = Literal['source', 'mask', 'phases', 'no_overlap', 'refractive_index']
+Options_rmul = Literal['intensity', 'amplitude', 'phase']
 
 Options_squares_nxm = Literal['amplitude', 'intensity', 'gray_levels']
 video_isovalue_Options = Literal['intensity', 'refractive_index']
 
 get_scalar_options = Literal['intensity', 'phase','field']
 get_vector_options = Literal['E', 'H', 'EH', 'fields', 'intensity', 'intensities', 'phases', 'poynting_vector', 
-                             'poynting_vector_averaged',  'poynting_total', 
-                             'energy_density', 'irradiance', 'stokes', 'params_ellipse']
+                             'poynting_vector_averaged',  'poynting_total',  'energy_density', 'irradiance', 
+                             'stokes', 'params_ellipse']